Move NullKeys out of their own package
