+++ finance/src/main/kotlin/net/corda/finance/contracts/asset/Cash.kt
@@ -6,7 +6,7 @@ import net.corda.finance.contracts.asset.cash.selection.CashSelectionH2Impl
 import net.corda.core.contracts.*
 import net.corda.core.contracts.Amount.Companion.sumOrThrow
 import net.corda.core.crypto.entropyToKeyPair
-import net.corda.core.crypto.testing.NULL_PARTY
+import net.corda.core.crypto.NullKeys.NULL_PARTY
 import net.corda.core.crypto.toBase58String
 import net.corda.core.identity.AbstractParty
 import net.corda.core.identity.Party
