+++ finance/src/test/kotlin/net/corda/finance/contracts/asset/ObligationTests.kt
@@ -3,7 +3,7 @@ package net.corda.finance.contracts.asset
 import net.corda.core.contracts.*
 import net.corda.core.crypto.SecureHash
 import net.corda.core.crypto.sha256
-import net.corda.core.crypto.testing.NULL_PARTY
+import net.corda.core.crypto.NullKeys.NULL_PARTY
 import net.corda.core.identity.AbstractParty
 import net.corda.core.identity.AnonymousParty
 import net.corda.core.transactions.TransactionBuilder
