+++ test-utils/src/main/kotlin/net/corda/testing/TestDSL.kt
@@ -2,8 +2,8 @@ package net.corda.testing
 
 import net.corda.core.contracts.*
 import net.corda.core.crypto.*
+import net.corda.core.crypto.NullKeys.NULL_SIGNATURE
 import net.corda.core.crypto.composite.CompositeKey
-import net.corda.core.crypto.testing.NULL_SIGNATURE
 import net.corda.core.identity.Party
 import net.corda.core.node.ServiceHub
 import net.corda.core.transactions.SignedTransaction
